demos/UC1_Legacy_Mainframe_Modernization/requirements.txt
# QBITEL - UC1 Legacy Mainframe Modernization Demo
# Production Requirements

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
python-multipart>=0.0.6
httpx>=0.26.0

# Authentication
PyJWT>=2.8.0
python-jose[cryptography]>=3.3.0

# Database
sqlalchemy>=2.0.25
aiosqlite>=0.19.0
asyncpg>=0.29.0  # PostgreSQL async driver

# Monitoring & Metrics
prometheus-client>=0.19.0
python-json-logger>=2.0.7

# LLM Integration (from main QBITEL)
openai>=1.12.0
anthropic>=0.18.0

# Utilities
pyyaml>=6.0.1
python-dotenv>=1.0.0
aiofiles>=23.2.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Development
black>=24.1.0
isort>=5.13.0
mypy>=1.8.0
ruff>=0.1.0
